updated some files to move code out of bot.py into where it belogns
