--- Read Me.txt
@@ -1,10 +1,0 @@
-Copyright (c) 2016, Situjuh Nazara (c7nazara@gmail.com),
-with Reserved Font Name GOBOLD. All rights reversed.
-
-License
-
-Personal use: FREE
-
-Commercial use: DONATE, mininal $10 USD via Situjuh's paypal account[c7nazara@gmail.com]. User can also buy the commercial license on https://www.creativefabrica.com/designer/situjuh/  
-
-If you have any request or question, contact me via email [c7nazara@gmail.com] or leave a comment on my website http://c7n1.com
